+++ src/javascript/crypto/e2e/extension/launcher.js
@@ -22,6 +22,7 @@
 
 goog.provide('e2e.ext.AppLauncher');
 goog.provide('e2e.ext.ExtensionLauncher');
+goog.provide('e2e.ext.yExtensionLauncher'); //@yahoo
 goog.provide('e2e.ext.Launcher');
 
 goog.require('e2e.ext.Preferences');
@@ -46,6 +47,7 @@ ext.Launcher = function(pgpContext, preferencesStorage) {
   /**
    * Whether the launcher was started correctly.
    * @type {boolean}
+   * @private
    */
   this.started_ = false;
 
@@ -68,6 +70,7 @@ ext.Launcher = function(pgpContext, preferencesStorage) {
    * The context API that the rest of the extension can use to communicate with
    * the PGP context.
    * @type {!ext.api.Api}
+   * @private
    */
   this.ctxApi_ = new ext.api.Api();
 };
@@ -264,5 +267,237 @@ ext.AppLauncher.prototype.createWindow = function(url, isForeground, callback) {
       callback);
 };
 
+
+
+/**
+ * Constructor to use in End-To-End Chrome extension.
+ * @param {!e2e.openpgp.Context} pgpContext The OpenPGP context to use.
+ * @param {!goog.storage.mechanism.IterableMechanism} preferencesStorage Storage
+ * mechanism for user preferences.
+ * @constructor
+ * @extends {ext.ExtensionLauncher}
+ */
+ext.yExtensionLauncher = function(pgpContext, preferencesStorage) {
+  /**
+   * The ID of the last used tab.
+   * @type {number}
+   * @private
+   */
+  this.lastTabId_ = window.NaN;
+
+  ext.yExtensionLauncher.base(this, 'constructor', pgpContext,
+      preferencesStorage);
+};
+goog.inherits(ext.yExtensionLauncher, ext.ExtensionLauncher);
+
+
+/** @override */
+ext.yExtensionLauncher.prototype.start = function(opt_passphrase) {
+
+  // All ymail tabs need to be reloaded for the e2ebind API to work
+  e2e.ext.utils.action.refreshYmail();
+
+  return goog.base(this, 'start', opt_passphrase).addCallback(function(){
+
+    // add message listener
+    chrome.runtime.onMessage.addListener(goog.bind(function(message, sender) {
+      this.proxyMessage(message, sender);
+    }, this));
+
+  }, this);
+};
+
+
+/**
+ * Stops the launcher. Called to lock keyring in extension/ui/prompt/prompt.js
+ */
+ext.yExtensionLauncher.prototype.stop = function() {
+  this.started_ = false;
+  // Unset the passphrase on the keyring
+  // TODO: add unsetKeyringPassphrase() into Context
+  this.getContext().keyring_ = null;
+  // Remoeve the API
+  this.ctxApi_.removeApi();
+  this.updatePassphraseWarning();
+  this.getActiveTab_(goog.bind(function(tabId) {
+    chrome.tabs.reload(tabId);
+  }, this));
+};
+
+
+/** @override */
+ext.yExtensionLauncher.prototype.showWelcomeScreen = function() {
+  var preferences = this.getPreferences();
+  if (preferences.isWelcomePageEnabled()) {
+    this.createWindow('setup.html', true, goog.nullFunction);
+    preferences.setWelcomePageEnabled(false);
+  }
+};
+
+
+/**
+ * TODO: @yahoo use WebsiteApi
+ * Sets the provided content into the element on the page that the user has
+ * selected.
+ * Note: This function might not work while debugging the extension.
+ * @param {string} content The content to write inside the selected element.
+ * @param {!Array.<string>} recipients The recipients of the message.
+ * @param {string} origin The web origin where the original message was created.
+ * @param {boolean} expectMoreUpdates True if more updates are expected. False
+ *     if this is the final update to the selected content.
+ * @param {!function(...)} callback The function to invoke once the content has
+ *     been updated.
+ * @param {string=} opt_subject The subject of the message if applicable.
+ * @export
+ */
+ext.yExtensionLauncher.prototype.updateSelectedContent =
+    function(content, recipients, origin, expectMoreUpdates,
+             callback, opt_subject) {
+  this.getActiveTab_(goog.bind(function(tabId) {
+    chrome.tabs.sendMessage(tabId, {
+      value: content,
+      response: true,
+      detach: !Boolean(expectMoreUpdates),
+      origin: origin,
+      recipients: recipients,
+      subject: opt_subject
+    });
+    callback();
+  }, this), true);
+};
+
+
+/**
+ * Retrieves the content that the user has selected.
+ * @param {!function(...)} callback The callback where the selected content will
+ *     be passed.
+ * @export
+ */
+ext.yExtensionLauncher.prototype.getSelectedContent = function(callback) {
+  this.getActiveTab_(goog.bind(function(tabId) {
+    chrome.tabs.sendMessage(tabId, {
+      editableElem: true,
+      enableLookingGlass: this.getPreferences().isLookingGlassEnabled(),
+      hasDraft: false
+    }, callback);
+  }, this), true);
+};
+
+
+/**
+ * Finds the current active tab.
+ * @param {!function(...)} callback The function to invoke once the active tab
+ *     is found.
+ * @param {boolean=} opt_runHelper Whether the helper script must be run first.
+ * @private
+ */
+ext.yExtensionLauncher.prototype.getActiveTab_ = function(callback, opt_runHelper) {
+  var runHelper = opt_runHelper || false;
+  chrome.tabs.query({
+    active: true,
+    currentWindow: true
+  }, goog.bind(function(tabs) {
+    var tab = tabs[0];
+    if (!goog.isDef(tab)) {
+      // NOTE(radi): In some operating systems (OSX, CrOS), the query will be
+      // executed against the window holding the browser action. In such
+      // situations we'll provide the last used tab.
+      callback(this.lastTabId_);
+      return;
+    } else {
+      this.lastTabId_ = tab.id;
+    }
+
+    // NOTE(yan): The helper script is executed automaticaly on ymail pages.
+    if (e2e.ext.utils.text.isYmailOrigin(tab.url) || !runHelper) {
+      callback(tab.id);
+    } else {
+      try {
+        chrome.tabs.executeScript(tab.id, {file: 'helper_binary.js'},
+                                  function() {
+                                    callback(tab.id);
+                                  });
+      } catch (e) {
+        // chrome-extension:// tabs throw an error. Ignore.
+        callback(tab.id);
+      }
+    }
+  }, this));
+};
+
+/**
+ * Proxies a message to the active tab.
+ * @param {messages.proxyMessage} incoming
+ * @param {!MessageSender} sender
+ */
+ext.yExtensionLauncher.prototype.proxyMessage = function(incoming, sender) {
+  if (incoming.proxy === true) {
+    var message = /** @type {messages.proxyMessage} */ (incoming);
+    this.getActiveTab_(goog.bind(function(tabId) {
+      // Execute the action, then forward the response to the correct tab.
+      var response = this.executeRequest_(message, tabId);
+      if (response) {
+        console.log('launcher proxying', response, tabId);
+        chrome.tabs.sendMessage(tabId, response);
+      }
+    }, this));
+  }
+};
+
+
+/**
+* Executes an action from a proxy request.
+* @param {messages.proxyMessage} args The message request
+* @param {number} tabId The ID of the active tab
+* @return {?(messages.proxyMessage|messages.BridgeMessageResponse|
+*           messages.GetSelectionRequest)}
+* @private
+*/
+ext.yExtensionLauncher.prototype.executeRequest_ = function(args, tabId) {
+  if (args.action === constants.Actions.GLASS_CLOSED ||
+      args.action === constants.Actions.SET_GLASS_SIZE) {
+    return /** @type {messages.proxyMessage} */ ({
+      content: args.content,
+      action: args.action
+    });
+  } else if (args.action === constants.Actions.SET_AND_SEND_DRAFT) {
+    var content = args.content;
+    return /** @type {messages.BridgeMessageResponse} */ ({
+      value: content.value,
+      response: content.response,
+      detach: content.detach,
+      origin: content.origin,
+      recipients: content.recipients,
+      subject: content.subject,
+      from: content.from
+    });
+  } else if (args.action === constants.Actions.GET_SELECTED_CONTENT) {
+    return /** @type {messages.GetSelectionRequest} */ ({
+      editableElem: true,
+      hasDraft: true,
+      enableLookingGlass: this.getPreferences().isLookingGlassEnabled()
+    });
+  } else if (args.action === constants.Actions.CHANGE_PAGEACTION) {
+    chrome.browserAction.setTitle({
+      tabId: tabId,
+      title: chrome.i18n.getMessage('composeGlassTitle')
+    });
+    chrome.browserAction.setIcon({
+      tabId: tabId,
+      path: 'images/yahoo/icon-128-green.png'
+    });
+    return null;
+  } else if (args.action === constants.Actions.RESET_PAGEACTION) {
+    chrome.browserAction.setTitle({
+      title: chrome.i18n.getMessage('extName'),
+      tabId: tabId
+    });
+    chrome.browserAction.setIcon({
+      tabId: tabId,
+      path: 'images/yahoo/icon-128.png'
+    });
+    return null;
+  }
+};
 
 });  // goog.scope
